+++ examples/attributes/server.py
@@ -1,14 +1,6 @@
 from __future__ import print_function
-import sys
 import Pyro4
 
-
-if sys.version_info < (3, 0):
-    input = raw_input
-
-print("Note that DOTTEDNAMES is a deprecated feature that will be removed in the next version.")
-dotted = input("enter value for DOTTEDNAMES config item: ").strip()
-Pyro4.config.DOTTEDNAMES = dotted in ("1", "true", "on", "yes")
 
 something = "Something"
 
@@ -47,9 +39,6 @@ class Thingy(object):
     def addSubValue(self, amount):
         self.sub.addToValue(amount)
 
-    def dottedNames(self):
-        return Pyro4.config.DOTTEDNAMES
-
     def printSomething(self):
         print("something:", something)
         return something
@@ -76,6 +65,5 @@ class Thingy(object):
 d = Pyro4.Daemon()
 uri = d.register(Thingy(), "example.attributes")
 print("server object uri:", uri)
-print("DOTTEDNAMES=", Pyro4.config.DOTTEDNAMES)
 print("attributes server running.")
 d.requestLoop()