+++ src/Pyro4/naming.py
@@ -113,8 +113,6 @@ class NameServerDaemon(core.Daemon):
     """Daemon that contains the Name Server."""
 
     def __init__(self, host=None, port=None, unixsocket=None, nathost=None, natport=None):
-        if Pyro4.config.DOTTEDNAMES:
-            raise PyroError("Name server won't start with DOTTEDNAMES enabled because of security reasons")
         if host is None:
             host = Pyro4.config.HOST
         if port is None: